Use long for TuringMachine::tapeToNumber

+++ TuringMachine.cpp
@@ -32,15 +32,15 @@ bool TuringMachine::simulate() {
 // then one after it, then two before it, then two after it, and so on.
 // We convert this interleaving into a natural number via binary with the smallest power on the left,
 // rather than the common convention of the smallest on the right
-int TuringMachine::tapeToNumber() {
-    int num = 0;
+long TuringMachine::tapeToNumber() {
+    long num = 0;
     for (int i = 0; i < tape.size(); i++) {
         if (i < startShift) {
             int placesLeft = startShift - i;
-            num += tape[i] << (placesLeft * 2 - 1);
+            num += (long)tape[i] << (placesLeft * 2 - 1);
         } else {
             int placesRight = i - startShift;
-            num += tape[i] << placesRight * 2;
+            num += (long)tape[i] << placesRight * 2;
         }
     }
     return num;
