+++ TuringMachine.h
@@ -23,7 +23,7 @@ public:
     std::vector<bool> tape = {false};
 
     bool simulate();
-    int tapeToNumber();
+    long tapeToNumber();
 };
 
 #endif //TURING_TURINGMACHINE_H
